Refactor package names for project consistency
All package names within the project were updated to maintain proper casing. The names have been changed from "habittrackerbackend" to "habitTrackerBackend", improving readability and adhering to Java naming conventions.
