+++ src/main/java/com/sa/habitTrackerBackend/HabitTrackerBackendApplication.java
@@ -1,0 +1,13 @@
+package com.sa.habitTrackerBackend;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HabitTrackerBackendApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(HabitTrackerBackendApplication.class, args);
+	}
+
+}
