+++ src/main/java/com/sa/habitTrackerBackend/configurations/auth/CustomUserDetails.java
@@ -1,0 +1,27 @@
+package com.sa.habitTrackerBackend.configurations.auth;
+
+import com.sa.habitTrackerBackend.dao.UserEntity;
+import lombok.Getter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+
+import java.util.Collection;
+import java.util.Objects;
+
+@Getter
+public class CustomUserDetails extends User {
+    private final transient UserEntity userEntity;
+
+    public CustomUserDetails(UserEntity user , Collection<? extends GrantedAuthority> authorities) {
+        super(user.getEmail(), user.getPassword(), authorities);
+        this.userEntity = user;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof CustomUserDetails that)) return false;
+        if (!super.equals(o)) return false;
+        return Objects.equals(userEntity, that.userEntity);
+    }
+}
