+++ src/main/java/com/sa/habitTrackerBackend/conteoller/TestController.java
@@ -1,0 +1,14 @@
+package com.sa.habitTrackerBackend.conteoller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/test")
+public class TestController {
+    @GetMapping
+    public String test() {
+        return "Test";
+    }
+}
