+++ src/main/java/com/sa/habitTrackerBackend/dto/user/LoginUserDto.java
@@ -1,0 +1,12 @@
+package com.sa.habitTrackerBackend.dto.user;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class LoginUserDto {
+    private String email;
+
+    private String password;
+}
