+++ src/main/java/com/sa/habitTrackerBackend/dto/user/UserDto.java
@@ -1,0 +1,12 @@
+package com.sa.habitTrackerBackend.dto.user;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class UserDto {
+    private String firstName;
+    private String lastName;
+    private String email;
+}
