+++ src/main/java/com/sa/habitTrackerBackend/dto/user/UserLoginResponseDto.java
@@ -1,0 +1,12 @@
+package com.sa.habitTrackerBackend.dto.user;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class UserLoginResponseDto {
+    private String token;
+
+    private long expiresIn;
+}
