+++ src/main/java/com/sa/habitTrackerBackend/service/user/UserService.java
@@ -1,0 +1,7 @@
+package com.sa.habitTrackerBackend.service.user;
+
+import com.sa.habitTrackerBackend.dao.UserEntity;
+
+public interface UserService {
+    UserEntity getUserByEmail(String email);
+}
