+++ src/main/java/com/sa/habitTrackerBackend/service/user/UserServiceImpl.java
@@ -1,0 +1,21 @@
+package com.sa.habitTrackerBackend.service.user;
+
+import com.sa.habitTrackerBackend.dao.UserEntity;
+import com.sa.habitTrackerBackend.repository.UserRepository;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserServiceImpl implements UserService{
+    private final UserRepository userRepository;
+
+    public UserServiceImpl(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+
+    @Override
+    public UserEntity getUserByEmail(String email) {
+        return userRepository.findByEmail(email).orElseThrow(() -> new UsernameNotFoundException(String.format("User with email %s not found", email)));
+    }
+}
