+++ src/main/java/com/sa/habitTrackerBackend/utils/enums/RolesEnum.java
@@ -1,0 +1,7 @@
+package com.sa.habitTrackerBackend.utils.enums;
+
+public enum RolesEnum {
+    SUPER_ADMIN,
+    ADMIN,
+    USER
+}
