--- src/main/java/com/sa/habittrackerbackend/HabitTrackerBackendApplication.java
@@ -1,13 +1,0 @@
-package com.sa.habittrackerbackend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HabitTrackerBackendApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(HabitTrackerBackendApplication.class, args);
-	}
-
-}
