--- src/main/java/com/sa/habittrackerbackend/commons/exception/AbstractCheckedException.java
@@ -1,22 +1,0 @@
-package com.sa.habittrackerbackend.commons.exception;
-
-import lombok.Getter;
-
-@Getter
-public abstract class AbstractCheckedException extends Exception {
-
-    protected final transient Object details;
-    protected final String code;
-
-    protected AbstractCheckedException(String message, Object details, String code) {
-        super(message);
-        this.details = details;
-        this.code = code;
-    }
-
-    protected AbstractCheckedException(Throwable cause, String message, Object details, String code) {
-        super(message, cause);
-        this.code = code;
-        this.details = details;
-    }
-}
