--- src/main/java/com/sa/habittrackerbackend/configurations/swagger/SwaggerConfig.java
@@ -1,22 +1,0 @@
-package com.sa.habittrackerbackend.configurations.swagger;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-    @Value("${info.app.version}")
-    private String appVersion;
-
-    @Bean
-    public OpenAPI customOpenAPI() {
-        return new OpenAPI()
-                .info(new Info()
-                        .title("Habit Tracker Backend API")
-                        .description("Habit Tracker is an application to record and track habits and activities.")
-                        .version(appVersion));
-    }
-}
