--- src/main/java/com/sa/habittrackerbackend/conteoller/TestController.java
@@ -1,14 +1,0 @@
-package com.sa.habittrackerbackend.conteoller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/test")
-public class TestController {
-    @GetMapping
-    public String test() {
-        return "Test";
-    }
-}
