--- src/main/java/com/sa/habittrackerbackend/dto/user/UserDto.java
@@ -1,12 +1,0 @@
-package com.sa.habittrackerbackend.dto.user;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class UserDto {
-    private String firstName;
-    private String lastName;
-    private String email;
-}
