--- src/main/java/com/sa/habittrackerbackend/dto/user/UserLoginResponseDto.java
@@ -1,12 +1,0 @@
-package com.sa.habittrackerbackend.dto.user;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class UserLoginResponseDto {
-    private String token;
-
-    private long expiresIn;
-}
