--- src/main/java/com/sa/habittrackerbackend/service/user/UserService.java
@@ -1,7 +1,0 @@
-package com.sa.habittrackerbackend.service.user;
-
-import com.sa.habittrackerbackend.dao.UserEntity;
-
-public interface UserService {
-    UserEntity getUserByEmail(String email);
-}
