--- src/main/java/com/sa/habittrackerbackend/service/user/UserServiceImpl.java
@@ -1,21 +1,0 @@
-package com.sa.habittrackerbackend.service.user;
-
-import com.sa.habittrackerbackend.dao.UserEntity;
-import com.sa.habittrackerbackend.repository.UserRepository;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserServiceImpl implements UserService{
-    private final UserRepository userRepository;
-
-    public UserServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-
-    @Override
-    public UserEntity getUserByEmail(String email) {
-        return userRepository.findByEmail(email).orElseThrow(() -> new UsernameNotFoundException(String.format("User with email %s not found", email)));
-    }
-}
