--- src/main/java/com/sa/habittrackerbackend/utils/enums/RolesEnum.java
@@ -1,7 +1,0 @@
-package com.sa.habittrackerbackend.utils.enums;
-
-public enum RolesEnum {
-    SUPER_ADMIN,
-    ADMIN,
-    USER
-}
